refactor(reporter): split to different files
